Change data type to long double for higher precision.
- Include decimal write in output file.

+++ src/include/matops.h
@@ -6,7 +6,7 @@ struct matrix
 {
     int rows;
     int cols;
-    double *data;
+    long double *data;
 };
 
 matrix *mat_mul(matrix *A, matrix *B);
